requirements.txt
appdirs==1.4.4
attrs==19.3.0
autopep8==1.5.3
black==19.10b0
click==7.1.2
pathspec==0.8.0
Pillow==7.2.0
pycodestyle==2.6.0
regex==2020.7.14
toml==0.10.1
typed-ast==1.4.1
